% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emotions}
\alias{emotions}
\title{Emotions}
\format{
A data frame with stopwords.
}
\source{
\url{https://github.com/peterdalle/svensktext}
}
\usage{
emotions
}
\description{
Swedish stopwords.
}
\keyword{datasets}
