% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{governmentorganizations}
\alias{governmentorganizations}
\title{Title}
\format{
data
}
\source{
\url{https://github.com/peterdalle/svensktext}
}
\usage{
governmentorganizations
}
\description{
Title
}
\keyword{datasets}
