% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_men_tilltalsnamn}
\alias{name_men_tilltalsnamn}
\title{Title}
\format{
data
}
\source{
\url{https://github.com/peterdalle/svensktext}
}
\usage{
name_men_tilltalsnamn
}
\description{
Title
}
\keyword{datasets}
