% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Stopwords}
\format{
A data frame with stopwords.

data
}
\source{
\url{https://github.com/peterdalle/svensktext}

\url{https://github.com/peterdalle/svensktext}
}
\usage{
stopwords

stopwords
}
\description{
Swedish stopwords.
}
\keyword{datasets}
