% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{svensktext}
\alias{svensktext}
\title{Svensk text}
\description{
This package contains Swedish text data formatted
to be convenient for text analysis.
}
\details{
The text repository can be found at
\href{https://github.com/peterdalle/svensktext}.
}
